# svc/home-api/openapi.yaml
openapi: 3.0.3
info:
  title: Home API
  version: 2.0.0
  description: >
    Frontend-facing Breathing House API for rooms, sensors, room history,
    alerts, and sensor gateway availability.


    Domain model: a room is an explicit, user-managed place (technical
    roomId = server-generated UUID, editable name and description). A sensor
    is a physical device (technical sensorId, editable displayName). Sensors are
    paired through the API or registered automatically when they send data.
    A sensor belongs to at most one room at a time and
    can be moved or unassigned. Every history record snapshots the room the
    sensor was assigned to when the record was ingested, so history stays
    under the original room after a move and records ingested while a sensor
    is unassigned are not returned by any room endpoint.


    Breaking changes from 1.x: rooms are no longer discovered from history;
    PATCH /rooms/{roomId} takes name/description instead of displayName;
    history items expose sensorId instead of deviceId; sensors and room
    assignment endpoints were added. The service evaluates configurable alert
    rules internally, exposes the resulting alerts through a read-only API
    (no write, acknowledge, or notification endpoints), and reports whether
    the sensor gateway is online from Kafka STATUS heartbeats.

tags:
  - name: Rooms
  - name: Sensors
  - name: History
  - name: Alerts

paths:
  /live:
    get:
      operationId: getLive
      summary: Check whether the service is alive
      responses:
        '200':
          description: The service is alive
          content:
            text/plain:
              schema:
                type: string
                example: OK
  /ready:
    get:
      operationId: getReady
      summary: Check whether PostgreSQL history and home_api tables are readable
      description: >
        Returns ready when environment, occupancy, alert, alert_state,
        gateway_heartbeat, room, and sensor tables are queryable. Gateway
        online/offline state is not part of readiness.
      responses:
        '200':
          description: Required database tables are queryable
          content:
            text/plain:
              schema:
                type: string
                example: READY
        '503':
          description: Database or required tables are unavailable
          content:
            text/plain:
              schema:
                type: string
                example: NOT_READY
  /api/v1/sensor-gateway/status:
    get:
      operationId: getSensorGatewayStatus
      summary: Report whether the sensor gateway is currently online
      description: >
        Online means a valid STATUS heartbeat for the configured gateway ID was
        received within SENSOR_GATEWAY_HEARTBEAT_TIMEOUT, using receivedAt.
        Missing or expired heartbeats return online=false with HTTP 200.
      responses:
        '200':
          description: Gateway availability domain state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayStatusResponse'
              examples:
                online:
                  value:
                    online: true
                offline:
                  value:
                    online: false
        '500':
          description: Database failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/rooms:
    post:
      operationId: createRoom
      tags: [Rooms]
      summary: Create a room
      description: >
        Creates an empty room. roomId is a server-generated UUID and cannot be
        chosen by the client. Names do not need to be unique.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoomRequest'
            example:
              name: Living Room
              description: Ground floor, south facing
      responses:
        '201':
          description: Room created
          headers:
            Location:
              description: URL of the created room
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
              example:
                roomId: 0b0d8f06-7d51-4a54-9c6f-3a0f8c5e0d11
                name: Living Room
                description: Ground floor, south facing
                sensorIds: []
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/DatabaseFailure'
    get:
      operationId: listRooms
      tags: [Rooms]
      summary: List rooms
      description: >
        Lists rooms stored in home_api.room (not rooms inferred from history),
        ordered by name then roomId. Each room includes the sensorIds that are
        assigned to it right now.
      responses:
        '200':
          description: Rooms with their current sensors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomsResponse'
        '500':
          $ref: '#/components/responses/DatabaseFailure'
  /api/v1/rooms/{roomId}:
    get:
      operationId: getRoom
      tags: [Rooms]
      summary: Get a room
      parameters:
        - $ref: '#/components/parameters/RoomId'
      responses:
        '200':
          description: The room with its current sensorIds
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/DatabaseFailure'
    patch:
      operationId: updateRoom
      tags: [Rooms]
      summary: Update a room's name and/or description
      description: >
        Partial update. At least one of name or description must be present;
        an empty patch is rejected. An omitted field is left unchanged; an
        explicit null description (or one that is blank after trimming) clears
        it. roomId never changes.
      parameters:
        - $ref: '#/components/parameters/RoomId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoomRequest'
            example:
              name: Salon
      responses:
        '200':
          description: Updated room
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/DatabaseFailure'
    delete:
      operationId: deleteRoom
      tags: [Rooms]
      summary: Delete a room
      description: >
        Hard-deletes the room. Assigned sensors become unassigned
        (`room_id` set to null). Active alerts for the room are resolved.
        Historical environment readings and occupancy events that reference
        the room id are retained.
      parameters:
        - $ref: '#/components/parameters/RoomId'
      responses:
        '204':
          description: Room deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/DatabaseFailure'
  /api/v1/rooms/{roomId}/sensors/{sensorId}:
    put:
      operationId: assignSensorToRoom
      tags: [Sensors]
      summary: Assign a sensor to a room
      description: >
        Atomically assigns the sensor to the room. A sensor is in at most one
        room, so if it was in another room it is moved. Moving or unassigning
        resolves the sensor's active alerts for the old room and re-evaluates
        composite room rules; alerts are never carried to the new room.
        Assigning a sensor to the room it is already in is a no-op.
      parameters:
        - $ref: '#/components/parameters/RoomId'
        - $ref: '#/components/parameters/SensorId'
      responses:
        '200':
          description: The updated sensor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sensor'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/DatabaseFailure'
    delete:
      operationId: unassignSensorFromRoom
      tags: [Sensors]
      summary: Unassign a sensor from a room
      description: >
        Leaves the sensor unassigned. Readings it sends while unassigned are
        stored without a room and do not appear on any room history endpoint
        or create room alerts.
      parameters:
        - $ref: '#/components/parameters/RoomId'
        - $ref: '#/components/parameters/SensorId'
      responses:
        '200':
          description: The updated (unassigned) sensor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sensor'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: The sensor is not currently assigned to this room
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                error: conflict
                message: Sensor 'air-1' is not currently assigned to room 'room-b'
        '500':
          $ref: '#/components/responses/DatabaseFailure'
  /api/v1/sensors:
    get:
      operationId: listSensors
      tags: [Sensors]
      summary: List known sensors
      description: >
        Lists paired sensors ordered by sensorId. types is derived from the
        distinct sensor_type / event_type values found in history for that
        sensor. A paired sensor with no history yet has an empty types list.
      responses:
        '200':
          description: Known sensors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SensorsResponse'
        '500':
          $ref: '#/components/responses/DatabaseFailure'
    post:
      operationId: pairSensor
      tags: [Sensors]
      summary: Pair a sensor
      description: >
        Adds the device to the known-sensor list. sensorId is the immutable
        identifier reported by the hardware. displayName defaults to sensorId
        and is not overwritten if the sensor is already paired (200).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PairSensorRequest'
            example:
              sensorId: air-11
      responses:
        '201':
          description: Sensor paired
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sensor'
        '200':
          description: Sensor was already paired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sensor'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/DatabaseFailure'
  /api/v1/sensors/{sensorId}:
    get:
      operationId: getSensor
      tags: [Sensors]
      summary: Get a discovered sensor
      parameters:
        - $ref: '#/components/parameters/SensorId'
      responses:
        '200':
          description: The sensor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sensor'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/DatabaseFailure'
    patch:
      operationId: updateSensor
      tags: [Sensors]
      summary: Rename a discovered sensor
      description: >
        Updates the presentation displayName only. sensorId never changes.
        Duplicate display names are allowed.
      parameters:
        - $ref: '#/components/parameters/SensorId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSensorRequest'
            example:
              displayName: Kitchen air
      responses:
        '200':
          description: Updated sensor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sensor'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/DatabaseFailure'
    delete:
      operationId: unpairSensor
      tags: [Sensors]
      summary: Unpair a sensor
      description: >
        Removes the device from the known-sensor list. History rows are kept.
        Active alerts for the sensor's current room are resolved. The sensor
        can be paired again later.
      parameters:
        - $ref: '#/components/parameters/SensorId'
      responses:
        '204':
          description: Sensor unpaired
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/DatabaseFailure'
  /api/v1/rooms/{roomId}/environment-readings:
    get:
      operationId: listEnvironmentReadings
      tags: [History]
      summary: List paginated environment readings for a room
      description: >
        Returns readings whose room snapshot is roomId. A reading stays under
        the room the sensor was in when it was ingested, even after the sensor
        is moved. Readings ingested while the sensor was unassigned are
        omitted. An unknown roomId returns an empty page.
      parameters:
        - $ref: '#/components/parameters/RoomId'
        - name: sensorType
          in: query
          schema:
            $ref: '#/components/schemas/SensorType'
        - name: sensorId
          in: query
          description: Optional filter for a single physical sensor
          schema:
            type: string
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Paginated environment readings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentReadingPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/DatabaseFailure'
  /api/v1/rooms/{roomId}/occupancy-events:
    get:
      operationId: listOccupancyEvents
      tags: [History]
      summary: List paginated occupancy events for a room
      description: >
        Same room-snapshot semantics as the environment readings endpoint.
      parameters:
        - $ref: '#/components/parameters/RoomId'
        - name: eventType
          in: query
          schema:
            $ref: '#/components/schemas/EventType'
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Paginated occupancy events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OccupancyEventPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/DatabaseFailure'
  /api/v1/alerts:
    get:
      operationId: listAlerts
      tags: [Alerts]
      summary: List paginated alerts
      description: >
        Read-only list of persisted alerts, newest first (triggeredAt DESC,
        id DESC). All filters are optional and combined with AND. The rule
        snapshot is only returned by the single-alert endpoint.
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/AlertStatus'
        - name: roomId
          in: query
          schema:
            type: string
        - name: sensorId
          in: query
          schema:
            type: string
        - name: severity
          in: query
          schema:
            $ref: '#/components/schemas/AlertSeverity'
        - name: from
          in: query
          description: Inclusive lower bound on triggeredAt (RFC3339); must not be after to
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: Inclusive upper bound on triggeredAt (RFC3339)
          schema:
            type: string
            format: date-time
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Paginated alerts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/DatabaseFailure'
  /api/v1/alerts/{id}:
    get:
      operationId: getAlert
      tags: [Alerts]
      summary: Get one alert including its rule snapshot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: The alert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/DatabaseFailure'

components:
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Room, sensor, or alert not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            error: not_found
            message: Room 'ghost' was not found
    DatabaseFailure:
      description: Database failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  parameters:
    RoomId:
      name: roomId
      in: path
      required: true
      description: Technical room identifier (server-generated UUID)
      schema:
        type: string
        minLength: 1
    SensorId:
      name: sensorId
      in: path
      required: true
      description: Technical sensor identifier reported by the device
      schema:
        type: string
        minLength: 1
    From:
      name: from
      in: query
      description: Inclusive lower bound on observedAt (RFC3339)
      schema:
        type: string
        format: date-time
    To:
      name: to
      in: query
      description: Inclusive upper bound on observedAt (RFC3339)
      schema:
        type: string
        format: date-time
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 100
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
  schemas:
    SensorType:
      type: string
      enum: [ROOM, AIR]
    EventType:
      type: string
      enum: [PRESENCE, OPENING]
    AlertStatus:
      type: string
      enum: [ACTIVE, RESOLVED]
    AlertSeverity:
      type: string
      enum: [INFO, WARNING, CRITICAL]
    Room:
      type: object
      required: [roomId, name, description, sensorIds]
      properties:
        roomId:
          type: string
          format: uuid
          description: Immutable technical identifier generated by the server
        name:
          type: string
          description: Mutable presentation name; not unique
          minLength: 1
          maxLength: 100
        description:
          type: string
          nullable: true
          maxLength: 500
        sensorIds:
          type: array
          description: Sensors currently assigned to this room, ordered by sensorId
          items:
            type: string
    CreateRoomRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          description: Trimmed before persistence; must be nonblank
          minLength: 1
          maxLength: 100
        description:
          type: string
          nullable: true
          description: Trimmed before persistence; blank is stored as null
          maxLength: 500
    UpdateRoomRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          nullable: true
          maxLength: 500
    RoomsResponse:
      type: object
      required: [rooms]
      properties:
        rooms:
          type: array
          items:
            $ref: '#/components/schemas/Room'
    SensorKind:
      type: string
      description: >
        Kind of data observed for a sensor in history. ROOM and AIR come from
        environment readings, PRESENCE and OPENING from occupancy events.
      enum: [ROOM, AIR, OPENING, PRESENCE]
    Sensor:
      type: object
      required: [sensorId, displayName, types, roomId]
      properties:
        sensorId:
          type: string
          description: Immutable technical identifier reported by the device
        displayName:
          type: string
          description: Mutable presentation name; defaults to sensorId; not unique
          minLength: 1
          maxLength: 100
        types:
          type: array
          description: Distinct kinds of data seen for this sensor in history, sorted
          items:
            $ref: '#/components/schemas/SensorKind'
        roomId:
          type: string
          nullable: true
          description: Room the sensor is currently assigned to; null when unassigned
    SensorsResponse:
      type: object
      required: [sensors]
      properties:
        sensors:
          type: array
          items:
            $ref: '#/components/schemas/Sensor'
    PairSensorRequest:
      type: object
      required: [sensorId]
      properties:
        sensorId:
          type: string
          description: Immutable technical identifier reported by the device
          minLength: 1
          maxLength: 200
        displayName:
          type: string
          description: Optional presentation name; defaults to sensorId on first pair
          minLength: 1
          maxLength: 100
    UpdateSensorRequest:
      type: object
      required: [displayName]
      properties:
        displayName:
          type: string
          description: Trimmed before persistence; must be nonblank
          minLength: 1
          maxLength: 100
    EnvironmentReading:
      type: object
      required:
        - id
        - roomId
        - sensorType
        - observedAt
        - receivedAt
        - ingestedAt
      properties:
        id:
          type: integer
          format: int64
        roomId:
          type: string
          description: Room the sensor was assigned to when this record was ingested
        sensorId:
          type: string
          nullable: true
          description: >
            Sensor that produced the record. Null for legacy records that never
            carried a sensor identity.
        sensorType:
          $ref: '#/components/schemas/SensorType'
        temperature:
          type: number
          nullable: true
        humidity:
          type: number
          nullable: true
        co2:
          type: number
          nullable: true
        light:
          type: number
          nullable: true
        lightLevel:
          type: string
          nullable: true
        observedAt:
          type: string
          format: date-time
        receivedAt:
          type: string
          format: date-time
        ingestedAt:
          type: string
          format: date-time
    OccupancyEvent:
      type: object
      required:
        - id
        - roomId
        - eventType
        - observedAt
        - receivedAt
        - ingestedAt
      properties:
        id:
          type: integer
          format: int64
        roomId:
          type: string
          description: Room the sensor was assigned to when this record was ingested
        sensorId:
          type: string
          nullable: true
          description: >
            Sensor that produced the record. Null for legacy records that never
            carried a sensor identity.
        eventType:
          $ref: '#/components/schemas/EventType'
        present:
          type: boolean
          nullable: true
        open:
          type: boolean
          nullable: true
        observedAt:
          type: string
          format: date-time
        receivedAt:
          type: string
          format: date-time
        ingestedAt:
          type: string
          format: date-time
    EnvironmentReadingPage:
      type: object
      required: [items, limit, offset, hasMore]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/EnvironmentReading'
        limit:
          type: integer
        offset:
          type: integer
        hasMore:
          type: boolean
    OccupancyEventPage:
      type: object
      required: [items, limit, offset, hasMore]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OccupancyEvent'
        limit:
          type: integer
        offset:
          type: integer
        hasMore:
          type: boolean
    Alert:
      type: object
      required:
        - id
        - ruleId
        - roomId
        - severity
        - status
        - message
        - triggeredAt
        - lastEvaluatedAt
      properties:
        id:
          type: integer
          format: int64
        ruleId:
          type: string
        roomId:
          type: string
        sensorId:
          type: string
          nullable: true
          description: Sensor that raised the alert; null for composite (room-level) alerts
        severity:
          $ref: '#/components/schemas/AlertSeverity'
        status:
          $ref: '#/components/schemas/AlertStatus'
        message:
          type: string
        triggerValue:
          type: string
          nullable: true
        triggeredAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        lastEvaluatedAt:
          type: string
          format: date-time
    AlertDetail:
      allOf:
        - $ref: '#/components/schemas/Alert'
        - type: object
          required: [ruleSnapshot]
          properties:
            ruleSnapshot:
              type: object
              additionalProperties: true
              description: Rule definition captured when the alert was raised
    AlertPage:
      type: object
      required: [items, limit, offset, hasMore]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Alert'
        limit:
          type: integer
        offset:
          type: integer
        hasMore:
          type: boolean
    GatewayStatusResponse:
      type: object
      required: [online]
      properties:
        online:
          type: boolean
          description: >
            True when the newest STATUS heartbeat for SENSOR_GATEWAY_ID has
            receivedAt within the configured timeout; otherwise false.
    ApiError:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string